% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_temps_gcms.R
\name{calc_temps_gcms}
\alias{calc_temps_gcms}
\title{calc_temps_gcms: calculate climate change metric(s) for emulator inputs}
\usage{
calc_temps_gcms(climate_dataset, mean_impute = FALSE)
}
\arguments{
\item{climate_dataset}{Climate data matrix object}

\item{mean_impute}{TRUE/FALSE to calculate ensemble mean of SSP to use as
forcings for simulations extended with impute_sims = "extend" (xxx check this)}
}
\value{
`calc_temps_gcms()` returns requested temperature change(s) for all
simulations in climate forcing file.
}
\description{
Calculate climate change metric(s) for emulator input from GCM forcings
}
