% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sle_anom.R
\name{calculate_sle_anom}
\alias{calculate_sle_anom}
\title{calculate_sle_anom: calculate sea level changes.}
\usage{
calculate_sle_anom()
}
\value{
`calculate_sle_anom()` returns ice_data with SL columns in
corrected units and relative to baseline year.
}
\description{
Calculate sea level changes (cm SLE) relative to a baseline year.
XXX This function will be obsolete if I standardise SLE file units and
baselines.
}
