% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{check_design}
\alias{check_design}
\title{check_design: check design matrix before sending to make_emu for emulation}
\usage{
check_design(designX)
}
\arguments{
\item{designX}{Dataset design matrix}
}
\value{
`check_design()` returns 1 if any potential problems found; 0 otherwise.
}
\description{
Test rank, conditioning, and other info and write to log file
Written with some input from Cursor AI.
}
