% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_LOO.R
\name{do_loo}
\alias{do_loo}
\title{do_LOO: calculate leave-one-out information.}
\usage{
do_loo(designX, responseF, forcingX, N_k = NA)
}
\arguments{
\item{designX}{Full dataset design}

\item{responseF}{Full dataset response}

\item{forcingX}{Full forcing timeseries (if SVD)}

\item{N_k}{Repeat the LOO for every N_k-th simulation (NA for all)}
}
\value{
`do_loo()` returns mean and sd of each test for each year.
}
\description{
Build emulator on all but one simulation and calculate difference with that
simulation. Repeat for list of all or a subset of the ensemble.
}
