% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_temps.R
\name{drop_temps}
\alias{drop_temps}
\title{drop_temps: check for highly correlated GSAT timeslices to drop}
\usage{
drop_temps(designX)
}
\arguments{
\item{designX}{Dataset design matrix}
}
\value{
`drop_temps()` returns subset of temp_list_names to drop, or NA.
}
\description{
Test GSAT timeslice correlation and return list to drop.
Written with some input from Cursor AI.
}
