% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator_predict.R
\name{emulator_predict}
\alias{emulator_predict}
\title{emulator_predict: Emulator predictions for design}
\usage{
emulator_predict(design_prior)
}
\arguments{
\item{design_prior}{Design for projections, i.e. sampled priors on all inputs.}
}
\value{
`emulator_predict()` returns emulator mean and sd values for design.
}
\description{
Predict for all points in design.
}
