% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator_uncertainty.R
\name{emulator_uncertainty}
\alias{emulator_uncertainty}
\title{emulator_uncertainty: sample from emulator uncertainties}
\usage{
emulator_uncertainty(myem_scen)
}
\arguments{
\item{myem_scen}{The emulator projections for a given scenario (contains
both the mean and variance).}
}
\description{
Generate final projections from the mean by sampling emulator
uncertainty once for each (i.e. same size dataset)
}
