% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glacier_cap.R
\name{get_glacier_cap}
\alias{get_glacier_cap}
\title{get_glacier_cap:}
\usage{
get_glacier_cap(region)
}
\arguments{
\item{region}{RGI region}
}
\value{
`get_glacier_cap()` returns maximum contribution (cm SLE)
}
\description{
Load max contribution for glacier region estimated by Farinotti et al.
}
