% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inputs.R
\name{get_inputs}
\alias{get_inputs}
\title{get_inputs: get input values for emulation}
\usage{
get_inputs(param_list)
}
\arguments{
\item{param_list}{Requested ice model inputs (continuous and categorical).}
}
\value{
`get_inputs()` returns requested (and imputed) columns from ice
data matrix.
}
\description{
Get input values for emulation and impute missing values.
XXX Impute all / factors etc in CSV instead?
}
