% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_design_to_pred.R
\name{load_design_to_pred}
\alias{load_design_to_pred}
\title{load_design_to_pred: construct design for emulator projections}
\usage{
load_design_to_pred(design_name)
}
\arguments{
\item{design_name}{Requested design: SSPs, uniform temperatures over original
ensemble ranges, or main effects.}
}
\value{
`load_design_to_pred()` returns requested designs for each list
member (SSP, or input for main effects).
}
\description{
Construct requested design for emulator projections by sampling inputs.
}
