% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_obs.R
\name{load_obs}
\alias{load_obs}
\title{load_obs: read observational data.}
\usage{
load_obs()
}
\value{
`load_obs()` returns annual observations with columns ["Year", "SLE",
"SLE_sd"]
}
\description{
Load IMBIE or Hugonnet et al. observations:
ICE SHEETS: IMBIE FOR AR6 1992-2020
Downloaded from https://ramadda.data.bas.ac.uk/repository/entry/show?entryid=77b64c55-7166-4a06-9def-2e400398e452
on 28/3/23 via doi https://doi.org/10.5285/77B64C55-7166-4A06-9DEF-2E400398E452
in IMBIE preprint https://essd.copernicus.org/preprints/essd-2022-261/essd-2022-261.pdf

GLACIERS: Hugonnet et al. (2021) regional files
https://www.sedoo.fr/theia-publication-products/?uuid=c428c5b9-df8f-4f86-9b75-e04c778e29b9
}
