% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sims.R
\name{load_sims}
\alias{load_sims}
\title{load_obs: climate forcings and land ice projections.}
\usage{
load_sims(variable, source = NA, region = NA)
}
\arguments{
\item{variable}{Variable: "climate" or "ice}

\item{source}{Ice source, i.e. value of i_s, if variable is ice}

\item{region}{Region - currently only used for GLA}
}
\value{
`load_sims()` returns CSV of dataset
}
\description{
Load current CSV file for climate, ice sheet or glacier simulations.
}
