% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_emu.R
\name{make_emu}
\alias{make_emu}
\title{make_emu: build emulator}
\usage{
make_emu(designX, responseF, r = NULL, thresh = 0.999)
}
\value{
`make_emu()` returns an emulator object to use.
}
\description{
Build emulators of principal components with RobustGaSP.
Code originally by Jonty Rougier.
}
