% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gcms.R
\name{match_gcms}
\alias{match_gcms}
\title{match_gcms: match ice simulations to climate forcing simulations}
\usage{
match_gcms(ice_data, temps_dataset)
}
\value{
`match_gcms()` returns requested temperature change(s) for each
simulation.
}
\description{
Look up climate simulation for each ice simulation by SSP and GCM, check
it exists, and calculate climate change for emulator input.
}
