% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sims.R
\name{match_sims}
\alias{match_sims}
\title{match_sims: match ice simulations to climate forcing simulations}
\usage{
match_sims()
}
\value{
`match_sims()` returns requested temperature change(s) for each
simulation.
}
\description{
Look up climate simulation for each ice simulation by SSP and GCM, check
it exists and is long enough, and calculate climate change for emulator input.
}
