% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayesian.R
\name{plot_bayesian}
\alias{plot_bayesian}
\title{plot_bayesian: plot prior and posterior distributions}
\usage{
plot_bayesian()
}
\description{
Plot prior and posterior density estimates from Bayesian
calibration.
}
