% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_designs.R
\name{plot_designs}
\alias{plot_designs}
\title{plot_designs: plot designs of ice simulations and emulator projections.}
\usage{
plot_designs(data_type, plot_level = 0)
}
\arguments{
\item{data_type}{Data type: "sims" for simulations; "prior" or "posterior" for
emulator projections before or after calibration.}

\item{plot_level}{Plot level: 0 for none, 1 for main, 2 for exhaustive}
}
\description{
Plot prior and posterior density estimates from Bayesian
calibration.
}
