% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{plot_distributions: plot distributions.}
\usage{
plot_distributions(data_type, plot_level = 0)
}
\arguments{
\item{data_type}{Data type: "sims" for simulations; "prior" or "posterior" for
emulator projections before or after calibration.}

\item{plot_level}{Plot level: 0 for none, 1 for main, 2 for exhaustive}
}
\description{
Plot distributions (histograms and density estimates). See
also plot_Bayesian.R.
}
