% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{plot_timeseries: timeseries plots.}
\usage{
plot_timeseries(data_type, plot_level = 0)
}
\arguments{
\item{data_type}{Data type: "sims" for simulations; "prior" or "posterior" for
emulator projections before or after calibration.}

\item{plot_level}{Plot level: 0 for none, 1 for main, 2 for exhaustive}
}
\description{
Timeseries figures of simulations and/or projections.
}
