% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_rows.R
\name{reorder_rows}
\alias{reorder_rows}
\title{Reorder the rows of a dataframe by factors' level}
\usage{
reorder_rows(X, frontLoad = FALSE)
}
\arguments{
\item{X}{A dataframe which may have factor variables.}

\item{frontLoad}{Logical, have roughly balanced numbers of each level at the start of the reorder.}
}
\value{
A permutation of the rows indices of `X`.
}
\description{
Find a permutation (stochastic) of the row indices of a dataframe with the property that rows `perm[1:n]` are roughly balanced according to the levels of the factors, for not-small `n`.
* If `frontLoad = FALSE` the levels in `perm[1:n]` are roughly representative of the proportions in the dataframe, for not-small `n`.
* If `frontLoad = TRUE` the levels in `perm[1:n]` are roughly balanced for not-small and not-large `n`: levels which are less common run out more quickly, and do not appear at the end of `perm`.
}
\details{
Internally, the factors are processed sequentially ordered by the KL divergence between the levels' proportions and the Uniform distribution, so that the more unbalanced factors (larger KL divergence) are prioritized.
}
\examples{
## Iris dataframe is balanced

p0 <- reorder_rows(iris)
show(iris[p0, ])

## add another factor to Dobson (1990) from ?lm

ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
X <- data.frame(weight = c(ctl, trt), group)
dow <- rep(c("Mon", "Tue", "Wed"), c(12, 5, 3))
X$dow <- factor(dow)

## representative
p1 <- reorder_rows(X) # frontLoad = FALSE
show(X[p1, ])

## balanced
p2 <- reorder_rows(X, frontLoad = TRUE)
show(X[p2, ])

}
