% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVDimpute.R
\name{SVDimpute}
\alias{SVDimpute}
\title{SVD Imputation Algorithm for Matrix Completion}
\usage{
SVDimpute(X, k = NULL, pmin = 0.99, reltol = 0.001, maxit = 20)
}
\arguments{
\item{X}{Numeric matrix containing missing values as `NA`.}

\item{k}{Desired rank.}

\item{pmin}{Minimum proportional variation, used to set `k` if `k = NULL`, see Details.}

\item{reltol}{Relative tolerance for exit condition, see Details.}

\item{maxit}{Maximum number of iterations, see Details.}
}
\value{
A matrix like `X` but with the `NA`s imputed, plus an attributes `k` and `niter`, the number of iterations.
}
\description{
Replace `NA` values in matrix `X` with imputed values.  Order `X` with time or equivalent down the rows, because the imputation is done row by row.

Although `k` can be specified directly, it is safer to set it indirectly using `pmin`.
}
\details{
The algorithm exits if:

1. There has been `maxit` iterations.
2. The norm of the change in the imputed values starts to rise.
3. The norm of the change in the imputed values falls below `reltol` times the norm of `X` (after scaling).

If `k` is not specified, the value of `k` is calculated from the SVD of `X` after it has been centered and scaled and `NA` values have been imputed using row means.  `k` is the smallest value for which the proportion of variation is at least `pmin`.
}
\references{
Olga Troyanskaya, Michael Cantor, Gavin Sherlock, Pat Brown, Trevor Hastie, Robert Tibshirani, David Botstein, Russ B. Altman, Missing value estimation methods for DNA microarrays, Bioinformatics, Volume 17, Issue 6, June 2001, Pages 520–525, <https://doi.org/10.1093/bioinformatics/17.6.520>.
}
