% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_glacier_cap.R
\name{weight_glacier_cap}
\alias{weight_glacier_cap}
\title{weight_glacier_cap:}
\usage{
weight_glacier_cap(proj_glac)
}
\value{
`weight_glacier_cap()` returns weighted projections
}
\description{
Weight posterior projections with max contribution for glacier region estimated by Farinotti et al.
}
