% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{check_design}
\alias{check_design}
\title{check_design: check design matrix before sending to make_emu for emulation}
\usage{
check_design(designX)
}
\arguments{
\item{designX}{Dataset design metric}

\item{strict}{Strict checking: fails on various criteria}
}
\value{
`check_design()` returns an emulator object to use.
}
\description{
Test rank, conditioning and GSAT correlation and write to log file
Written with some input from Cursor AI.
}
