% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_calibration.R
\name{do_calibration}
\alias{do_calibration}
\title{do_calibration: calculate calibration information.}
\usage{
do_calibration(distances, calib_type)
}
\arguments{
\item{distances}{Distance of each historical emulator prediction from the
observations}

\item{calib_type}{history_matching or Bayesian}
}
\value{
`do_calibration()` returns index of NROY ensemble members if
history matching or weights for each ensemble member if Bayesian calibration.
}
\description{
Use distance of each emulator prediction over the historical period from
observations to calculate whether it is implausible/NROY (history matching),
or calculate weights using a Gaussian likelihood (Bayesian calibration).
}
