% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_climate.R
\name{impute_climate}
\alias{impute_climate}
\title{impute_climate: fill missing data in climate forcings}
\usage{
impute_climate(climate_dataset, construct_fixed = FALSE)
}
\arguments{
\item{climate_dataset}{Dataset}

\item{construct_fixed}{Set to TRUE to construct post-2100 forcings for GIS}
}
\value{
`impute_climate()` returns climate dataset
}
\description{
Imputes missing 2100 and 2300 values if 2099 and 2299 values are there
Constructs fixed climate for GIS post-2100 fixed_climate = 2100 simulations
Drops columns that are not needed
}
