% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot_scatter: plot scatter (X-Y) figures.}
\usage{
plot_scatter(data_type, plot_level = 0)
}
\arguments{
\item{data_type}{Data type: "sims" for simulations; "prior" or "posterior" for
emulator projections before or after calibration.}

\item{plot_level}{Plot level: 0 for none, 1 for main, 2 for exhaustive}
}
\description{
Plot scatter and X-Y plots, including main effects.
}
