% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sims.R
\name{select_sims}
\alias{select_sims}
\title{select_sims: select land ice simulations}
\usage{
select_sims(select_type)
}
\arguments{
\item{select_type}{Type of selection: main or history_match}
}
\value{
select_sims returns a matrix ice_data which is a subset of the original.
}
\description{
Select land ice simulations from ice data file: ice source, region, model(s),
sufficient length, and further ice-source-specific selections.
}
