% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_outputs.R
\name{write_outputs}
\alias{write_outputs}
\title{write_outputs: write CSVs of projections.}
\usage{
write_outputs(write_mean)
}
\arguments{
\item{write_mean}{TRUE/FALSE to write mean projections, as well as the
final/full projections with emulator uncertainties.}
}
\description{
Write projections to CSV: just final (i.e. with emulator
uncertainties), or optionally also the mean.
}
